+++ messaging/src/main/java/org/frankframework/extensions/mqtt/MqttSender.java
@@ -38,6 +38,8 @@ import org.frankframework.stream.Message;
  *
  * Links to <a href="https://www.eclipse.org/paho/files/javadoc" target="_blank">https://www.eclipse.org/paho/files/javadoc</a> are opened in a new window/tab because the response from eclipse.org contains header X-Frame-Options:SAMEORIGIN which will make the browser refuse to open the link inside this frame.
  *
+ * {@inheritDoc}
+ *
  * @author Niels Meijer
  */
 
@@ -62,14 +64,10 @@ public class MqttSender extends MqttFacade implements ISenderWithParameters {
 	}
 
 	@Override
-	public void start() {
-		super.start();
-	}
+	public void start() {}
 
 	@Override
-	public void stop() {
-		super.stop();
-	}
+	public void stop() {}
 
 	@Override
 	public void addParameter(IParameter p) {
@@ -87,16 +85,12 @@ public class MqttSender extends MqttFacade implements ISenderWithParameters {
 	@Override
 	public @Nonnull SenderResult sendMessage(@Nonnull Message message, @Nonnull PipeLineSession session) throws SenderException, TimeoutException {
 		try {
-			if (!client.isConnected()) {
-				super.start();
-			}
-
 			String topic = getTopic();
 			if (getParameterList() != null) {
 				IParameter topicParameter = getParameterList().findParameter(TOPIC_PARAMETER_NAME);
 
 				if (topicParameter != null) {
-					topic = topicParameter.getValue();
+					topic = topicParameter.getValue(null, message, session, false).toString();
 				}
 			}